JSRPC: Rename: cloudflare:entrypoints -> cloudflare:workers
