+++ src/cloudflare/internal/workers.d.ts
