+++ src/cloudflare/workers.ts
@@ -5,7 +5,7 @@
 // TODO(cleanup): C++ built-in modules do not yet support named exports, so we must define this
 //   wrapper module that simply re-exports the classes from the built-in module.
 
-import entrypoints from 'cloudflare-internal:entrypoints';
+import entrypoints from 'cloudflare-internal:workers';
 
 export const WorkerEntrypoint = entrypoints.WorkerEntrypoint;
 export const DurableObject = entrypoints.DurableObject;
